Avanzamento sviluppi

+++ src/main/java/it/iad/demofabrick/model/Amount.java
@@ -1,6 +1,7 @@
 package it.iad.demofabrick.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import lombok.Data;
 
@@ -8,10 +9,15 @@ import lombok.Data;
 public class Amount implements Serializable{
 	private static final long serialVersionUID = 4276383092713752396L;
 	
-	private int debtorAmount;
+	private BigDecimal debtorAmount;
+	
 	private String debtorCurrency;
-	private int creditorAmount;
+	
+	private BigDecimal creditorAmount;
+	
 	private String creditorCurrency;
+	
 	private String creditorCurrencyDate;
-	private int exchangeRate;
+	
+	private double exchangeRate;
 }
